Add data layer util + update layers to use util.
