+++ client/src/util/data_layer_util.ts
@@ -1,0 +1,30 @@
+/**
+ * Converts list of legend pairs (alternating numerical key, string hex color
+ * value) to a map of legend key : color hex.
+ * @param colorMapping the list of legend pairs
+ *
+ * Example:
+ * [0, '#9fcd7c', 10000, '#f7e5af'] => {'0' :  '#9fcd7c', '10000', '#f7e5af'}
+ */
+export function colorMapToBuckets(colorMapping: any[]): Map<string, string> {
+  const listOfLegendPairs: [string, string][] = [];
+  for (let index = 0; index < colorMapping.length; index++) {
+    if (index % 2 != 0) {
+      listOfLegendPairs.push([colorMapping[index - 1].toString(), colorMapping[index]]);
+    }
+  }
+  return new Map(listOfLegendPairs);
+}
+
+const LOCALHOST = 'localhost';
+
+/**
+ * Hostname of tileserver to be passed to mapbox.
+ *
+ * Returns location.hostname unless app is being run locally.
+ */
+export function tileServerHost() {
+  return location.hostname == LOCALHOST
+    ? process.env.VUE_APP_DEFAULT_TILESERVER_HOST
+    : location.hostname;
+}
